requirements.txt
aiohappyeyeballs==2.6.1

aiohttp==3.13.3

aiohttp-retry==2.9.1

aiosignal==1.4.0

altair==5.5.0

amqp==5.3.1

asgiref==3.8.1

attrs==25.3.0

billiard==4.2.1

blinker==1.9.0

cachetools==5.5.2

celery==5.4.0

certifi==2026.1.4

charset-normalizer==3.4.2

click==8.1.8

click-didyoumean==0.3.1

click-plugins==1.1.1

click-repl==0.3.0

colorama==0.4.6

contourpy==1.3.1

cron-descriptor==1.4.5

cx_Oracle==8.3.0

cycler==0.12.1

DateTime==5.5

distlib==0.3.9

dj-database-url==3.1.0

Django==5.1.6

django-browser-reload==1.21.0

django-celery-beat==2.7.0

django-cleanup==9.0.0

django-cors-headers==4.9.0

django-environ==0.12.0

django-phonenumber-field==8.4.0

django-ratelimit==4.1.0

django-storages==1.14.6

django-tailwind==4.4.2

django-timezone-field==7.1

djangorestframework==3.16.1

djangorestframework_simplejwt==5.5.1

dotenv==0.9.9

env==0.1.0

filelock==3.18.0

Flask==3.1.0

flask-cors==6.0.0

fonttools==4.56.0

frozenlist==1.8.0

gitdb==4.0.12

GitPython==3.1.44

google-api-core==2.29.0

google-api-python-client==2.188.0

google-auth==2.47.0

google-auth-httplib2==0.3.0

google-cloud-core==2.5.0

google-cloud-storage==3.8.0

google-crc32c==1.8.0

google-resumable-media==2.8.0

googleapis-common-protos==1.72.0

gunicorn==23.0.0

httplib2==0.31.1

idna==3.10

itsdangerous==2.2.0

Jinja2==3.1.5

joblib==1.5.1

jsonschema==4.23.0

jsonschema-specifications==2025.4.1

kiwisolver==1.4.8

kombu==5.4.2

MarkupSafe==3.0.2

matplotlib==3.10.1

multidict==6.7.0

myenv==201902.7

mysql==0.0.3

mysql-connector==2.2.9

mysql-connector-python==9.2.0

mysqlclient==2.2.7

narwhals==1.40.0

node==1.2.3

numpy==2.2.3

odict==1.9.0

packaging==24.2

pandas==2.2.3

phonenumberslite==9.0.22

pillow==11.1.0

platformdirs==4.3.7

plumber==1.7

prompt_toolkit==3.0.50

propcache==0.4.1

proto-plus==1.27.0

protobuf==6.31.0

psycopg2-binary==2.9.11

pyarrow==20.0.0

pyasn1==0.6.2

pyasn1_modules==0.4.2

pydeck==0.9.1

PyJWT==2.10.1

pyparsing==3.2.1

pytailwindcss==0.3.0

python-crontab==3.2.0

python-dateutil==2.9.0.post0

python-dotenv==1.2.1

pytz==2025.1

qrcode==8.2

referencing==0.36.2

requests==2.32.3

rpds-py==0.25.1

rsa==4.9.1

scikit-learn==1.6.1

scipy==1.15.3

setuptools==80.9.0

six==1.17.0

smmap==5.0.2

sqlparse==0.5.3

streamlit==1.45.1

tailwind==3.1.5b0

tenacity==9.1.2

threadpoolctl==3.6.0

toml==0.10.2

tornado==6.5.1

twilio==9.9.1

typing_extensions==4.13.2

tzdata==2025.1

uritemplate==4.2.0

urllib3==2.4.0

vine==5.1.0

virtualenv==20.29.3

VirtualEnvironment==0.1.2

watchdog==6.0.0

wcwidth==0.2.13

Werkzeug==3.1.3

wheel==0.45.1

whitenoise==6.11.0

yarl==1.22.0

zope.component==7.0

zope.deferredimport==6.0

zope.deprecation==6.0

zope.event==6.1

zope.hookable==8.2

zope.interface==7.2

zope.lifecycleevent==6.0

zope.proxy==7.1
boto3
xhtml2pdf
djangorestframework-simplejwt
djangorestframework
requests
oracledb